requirements.txt
python-dotenv
openai
pymongo
Pillow
git+https://github.com/abraham-ai/eden-sdk-py